refactor: improve project structure
