+++ src/api/firebase.ts
@@ -10,7 +10,7 @@ import {
 } from 'firebase/firestore';
 import { HighscoreEntry } from '../types/game';
 
-// Config to connect to firebase
+// Config to connect to Firebase
 const firebaseConfig = {
     apiKey: 'AIzaSyAwtGacFb6f5mA4Bjyu8KiDOM-QrHI7zMs',
     authDomain: 'greenk-scoreboard.firebaseapp.com',
@@ -26,7 +26,7 @@ const dbRef = getFirestore();
 const scoresRef = collection(dbRef, 'scores');
 
 /**
- * Fetches scores from firestore, limiting the results to the 10 higest scores
+ * Fetches scores from Firestore, limiting the results to the 10 higest scores
  * @returns the 10 highest scores from the database, or an empty list on error
  */
 export async function fetchHighscores(): Promise<HighscoreEntry[]> {
@@ -46,7 +46,7 @@ export async function fetchHighscores(): Promise<HighscoreEntry[]> {
 }
 
 /**
- * Adds a highscore to firestore for storage
+ * Adds a highscore to Firestore for storage
  * @param highscore the highscore to store
  */
 export async function addHighscore(highscore: HighscoreEntry) {
