+++ src/components/funfacts/funfact-card/index.tsx
