+++ src/components/solarpanel/SolarPanelComponent.css
