+++ src/components/solarpanel/SolarPanelComponent.tsx
