+++ src/components/solarpanel/index.tsx
