+++ src/components/stats/Diagram.tsx
