+++ src/components/stats/DiagramC.tsx
