+++ src/pages/home/Home.tsx
@@ -2,9 +2,9 @@ import { Box, Grid, GridItem, Text } from '@chakra-ui/react';
 import React from 'react';
 import GreenkLogo from '../../assets/greenk-logo.png';
 import NidarvollVideo from '../../assets/Nidarvoll-video.mp4';
-import FunfactSlideshow from '../../components/funfact-slideshow';
-import Scoreboard from '../../components/Scoreboard';
-import SolarPanelComponent from '../../components/solar-panel-component';
+import FunfactSlideshow from '../../components/funfacts/funfact-slideshow';
+import Scoreboard from '../../components/game/Scoreboard';
+import SolarPanelComponent from '../../components/solarpanel';
 
 export interface HomePropsI {}
